requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
django-environ==0.4.5
pytz==2019.3
sqlparse==0.3.0
psycopg2>=2.7,<3.0
# psycopg2
pysqlite3
# uwsgi
environ